Fix call to ls_ref_attr

+++ src/investigate.cpp
@@ -131,7 +131,7 @@ static void foreach_select_investigation(GtkIconView *icon_view,
 
   GdkPixbufLoader *pix_loader = gdk_pixbuf_loader_new();
   void *diamond_data;
-  unsigned int len;
+  size_t len;
   ls_ref_attr(newobj, "", &len, (unsigned char **) &diamond_data);
 
   // loop to work around crash
